+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Expect active development and potentially significant breaking changes in the `0
 ### vNEXT
 - Added support for query composition through fragments [Issue #338](https://github.com/apollostack/apollo-client/issues/338) and [PR #343](https://github.com/apollostack/apollo-client/pull/343)
 - **Add `fetchMore` method for for infinite scroll in pagination**
-  - This is done by adding to the watched query a `quietFields` array
+  - This is done by adding to the watched query a `quietArguments` array
     that lists all parameters used for pagination that will not be re-inflated
     in the `itemDataId` in the store.
   - `fetchMore` sets a `forceFetch` alongside a `fetchMore` in the new query